+++ packages/solid-1.0/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @solana-wallets/solid-1.0
 
+## 0.11.2
+
+### Patch Changes
+
+- @solana-wallets/core-1.0@0.8.5
+
 ## 0.11.1
 
 ### Patch Changes